requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
audio-separator==0.14.3
audioread==3.0.1
bitarray==3.7.1
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.4
decorator==5.2.1
diffusers==0.36.0
edge-tts==7.2.3
einops==0.8.1
einx==0.3.0
exceptiongroup==1.3.0
fairseq==0.12.2
faiss-cpu==1.7.3
fastapi==0.117.1
ffmpeg-python==0.2.0
ffmpy==0.6.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
frozendict==2.4.7
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
gradio==5.47.2
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.1
humanfriendly==10.0
hydra-core==1.0.7
idna==3.10
importlib_metadata==8.7.1
infer_rvc_python==1.1.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
librosa==0.9.1
llvmlite==0.45.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
networkx==3.4.2
ninja==1.13.0
noisereduce==3.0.3
numba==0.62.0
numpy==1.23.5
omegaconf @ git+https://huggingface.co/datasets/soiz1/omegaconf@5ddebad2589b6c34fc8c47d8a31029a45181a546
onnx==1.19.0
onnx2torch==1.5.15
onnxruntime-gpu==1.23.0
optimum==2.1.0
optimum-quanto==0.2.7
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pedalboard==0.9.17
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
portalocker==3.2.0
praat-parselmouth==0.4.6
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pyworld==0.3.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
resampy==0.4.3
runpod==1.7.5
rich==14.1.0
ruff==0.13.2
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.6.2
samplerate==0.2.1
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3
torchcrepe==0.0.20
tqdm==4.67.1
transformers==4.56.2
typeguard==4.2.0
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
diffusers